--- a/includes/bootstrap.php
+++ b/includes/bootstrap.php
@@ -56,6 +56,7 @@ class HRW_Plugin_Bootstrap
 			'class-hrw-data-merger.php',
 			'class-hrw-api-cache.php',
 			'class-hrw-json-timing.php',
+			'class-hrw-wordpress-timing.php',
 		];
 
 		foreach ($classes as $class_file) {
@@ -83,6 +84,11 @@ class HRW_Plugin_Bootstrap
 		// Initialize JSON timing to measure serialization performance
 		if (class_exists('HRW_JSON_Timing')) {
 			HRW_JSON_Timing::init();
+		}
+
+		// Initialize WordPress timing to measure WordPress lifecycle overhead
+		if (class_exists('HRW_WordPress_Timing')) {
+			HRW_WordPress_Timing::init();
 		}
 	}