… .DS_Store
… assets/.DS_Store
… includes/class-hrw-json-timing.php
… includes/class-hrw-response-headers.php
… includes/class-hrw-wordpress-timing.php